Trim documentation across src/, move TiDE-only modules, correct devenv comments
Three passes over documentation and structure. No behaviour changes.

Module layout

artifact and predict moved from models/ to models/tide/. Both are TiDE-only and
always were: ModelState holds a TideModel<NdArray>, the artifact S3 prefix is
hardcoded to models/tide/, the loader reads tide_parameters.json, and the unsafe
Sync impl exists because of TideModel specifically. Sitting a level up implied a
seam between "the artifact system" and a particular model that does not exist.

models/ is now the list of models, currently one. If a second arrives,
predict.rs already shows where the real seam is: consolidate_data,
filter_equity_bars, validate_predictions, insert_predictions, and
load_predictions_between have no model in them, while filter_to_trained_tickers
and generate_predictions do.

Documentation

653 comment lines removed and 388 rewritten, a net reduction of 265 across 44
files. Removed was of three kinds.

Stale references to things that no longer exist: 23 mentions of a Python
trainer, when the repository contains no Python; an inference/ module describing
a split that was merged; a `train` feature gate, when the crate has no features;
and pointers to memory notes and to .scratchpad files, which are gitignored and
so unreadable to anyone holding only the repository. Where such a reference
carried a real rule -- the Monday=1 weekday convention, quantile monotonic
sorting, inclusive thresholds, the lowercase-ticker collision -- the rule was
kept and only the attribution dropped.

Content derivable from the code beneath it: lib.rs re-describing six modules
that each document themselves, batch.rs listing a flattening order that is the
next ten lines, config.rs noting that fields are private.

Restatement: most long blocks made their point and then made it twice more.

Five factual errors went with them.

  - drift.rs claimed drift was "recorded in model_runs". There is no model_runs
    table; the trainer only warns and infos.
  - observability.rs named the services "data", "inference", and "portfolio",
    which was the three-process split deleted in the rebuild. Callers pass
    "fund", "dashboard", "tide-model-trainer".
  - observability.rs justified init_tracing_file_only twice by a "dashboard TUI"
    and its "alternate screen buffer". There is no TUI: no ratatui, no
    crossterm, and the dashboard is an axum web server.
  - export.rs said the trainer fetches its own data from Alpaca. It fetches from
    Massive.
  - seed_equity_bars.rs said the model needs seventy sessions. It needs forty --
    window_frame skips any ticker shorter than input_length + output_length.

Kept: anything a reader cannot recover from the code. The drain-timeout
arithmetic, why subscribing precedes recovery, why both pair legs are closed
even after an error, the sargable-predicate reasoning in eastern_day_bounds, the
signed-versus-magnitude correlation argument, and the crossed-book midpoint
hazard. Short module headers were added to config.rs, data.rs, and model.rs,
which had none.

devenv.nix

database:backup claimed to run automatically via pg_cron at 22:00 UTC. No such
job exists -- schema.sql schedules six, none a backup -- and pg_cron runs SQL
inside the database, so it cannot invoke pg_dump as written. The comment now
says it is manual and distinguishes it from the nightly parquet export the
market data sync does chain: only the pg_dump can restore equity_pairs, which
nothing else reconstructs.

data:seed carried the same seventy-session error as seed_equity_bars.rs.

ensure-test-services probed pg_isready with its defaults rather than the
endpoint the suite connects to. tests/common/mod.rs reads
TEST_DATABASE_URL_BASE, so the probe now derives host and port from the same
value and names the endpoint it tried when it fails. A data directory records
its port in postgresql.conf, so a profile initialised while 5432 was busy keeps
the shifted port and the old probe would loop for a minute against nothing.

Verified with devenv tasks run checks:rust and checks:base.
