+++ src/bin/fund.rs
@@ -31,22 +31,17 @@ const LISTENER_RECONNECT_DELAY: Duration = Duration::from_secs(5);
 
 /// How long to wait for running handlers after shutdown is requested.
 ///
-/// The handlers that matter here are the ones with side effects part-applied: a liquidation between
-/// two broker orders, an export between S3 and the purge, or any handler that has not yet written
-/// its terminal event. Dropping one of those mid-sequence leaves a `_requested` row with no outcome
-/// while the process logs a clean stop.
+/// What matters is a handler with side effects part-applied: a liquidation between two broker
+/// orders, an export between S3 and the purge, or one that has not yet written its terminal event.
+/// Dropping one leaves a `_requested` row with no outcome while the process logs a clean stop.
 ///
-/// Derived from the work rather than picked as a round number. The longest thing this can be waiting
-/// on is one pair being opened, which is two legs at
+/// Derived from the work. The longest wait is one pair opening, two legs at
 /// [`fund::portfolio::execute::FILL_TIMEOUT`] each — sixty seconds. Seventy-five leaves margin for
-/// the terminal event write, and stays well inside the 120 seconds process-compose allows before it
-/// escalates.
+/// the terminal event write and stays inside the 120 process-compose allows before escalating.
 ///
-/// That bound only holds because the evaluation pass stops *starting* pairs once shutdown is
-/// requested; see `entries_abandoned` in [`fund::portfolio::evaluate`]. Without that check the pass
-/// would work through every approved candidate sequentially and no fixed timeout would cover it —
-/// the previous sixty seconds was exactly one pair's worst case with nothing left over, while its
-/// own comment claimed it was "longer than any handler should take".
+/// The bound holds only because the evaluation pass stops *starting* pairs once shutdown is
+/// requested; see `entries_abandoned` in [`fund::portfolio::evaluate`]. Without that the pass works
+/// through every approved candidate and no fixed timeout covers it.
 const HANDLER_DRAIN_TIMEOUT: Duration = Duration::from_secs(75);
 
 #[tokio::main]
@@ -137,23 +132,17 @@ async fn listen(
         info!("Listening for events");
 
         // Recovery runs *after* subscribing and *before* the first `recv`, on every connection
-        // rather than once at startup. Both halves of that matter.
+        // rather than once at startup.
         //
         // After subscribing, because PostgreSQL delivers a NOTIFY only to sessions already
-        // listening — it is never persisted. Scanning first and subscribing second leaves a window
-        // in which a request is written, notified to nobody, and then missed by the scan that
-        // already ran. The row sits at `_requested` with no terminal outcome and nothing looks
-        // again until the next restart; for a liquidation that means positions held overnight.
+        // listening and never persists it. Scanning first leaves a window where a request is
+        // written, notified to nobody, and missed by the scan that already ran — the row sits at
+        // `_requested` until the next restart, which for a liquidation means positions held
+        // overnight. The same window reopens on every reconnect, and that one recurs unattended.
         //
-        // On every connection, because the same window reopens on every reconnect — and that one
-        // recurs unattended. Anything notified while the connection was down was addressed to a
-        // session that no longer exists.
-        //
-        // Subscribing first cannot double-dispatch. The subscription is passive: notifications
-        // queue on the connection until `recv` is called, which does not happen until this returns.
-        // A request that recovery replays and the queue also carries arrives twice, and `claim` --
-        // a mutex-guarded set insert -- admits exactly one of them. Re-scanning on every reconnect
-        // is likewise cheap and safe: the query is bounded to today's Eastern date, and
+        // Subscribing first cannot double-dispatch: the subscription is passive, so notifications
+        // queue until `recv` is called. A request both replayed and queued arrives twice, and
+        // `claim` admits one. Re-scanning is bounded to today's Eastern date, and
         // `Command::PortfolioEvaluation` is `Recovery::Skip`, so a reconnect cannot replay a pass
         // whose inputs have gone stale.
         handlers::recover(Arc::clone(&state)).await;
@@ -161,16 +150,11 @@ async fn listen(
             tokio::select! {
                 () = shutdown.cancelled() => return,
                 // `try_recv`, not `recv`. `recv` loops over `try_recv` internally and returns only
-                // notifications, so a dropped connection is invisible to the caller — sqlx
-                // reconnects, re-subscribes, and carries on, and its own documentation is explicit
-                // that "any received while the connection was lost will not be returned". A
-                // reconnect handled that transparently is a silent hole: the request row exists,
-                // nothing dispatched it, and no log line was ever written. Measured, not assumed —
-                // terminating the listener's backend and inserting a request in the gap left it at
-                // `_requested` with no terminal outcome and not one line in the log.
-                //
-                // `try_recv` surfaces that boundary as `Ok(None)`, which is the only place the
-                // rescan can be hung.
+                // notifications, so a dropped connection is invisible to the caller: sqlx
+                // reconnects and carries on, and its documentation is explicit that anything
+                // received while the connection was lost is not returned. That is a silent hole —
+                // the request row exists, nothing dispatched it, no log line. `try_recv` surfaces
+                // the boundary as `Ok(None)`, the only place the rescan can be hung.
                 received = listener.try_recv() => match received {
                     Ok(Some(notification)) => {
                         // Reap anything already finished so the set does not grow across a session.
@@ -178,19 +162,15 @@ async fn listen(
                         dispatch(&state, handlers_in_flight, notification.payload()).await
                     }
                     Ok(None) => {
-                        // sqlx has already reconnected and re-subscribed by the time this returns,
-                        // so there is nothing to rebuild — only the gap to close. Anything notified
-                        // while the connection was down is gone, and the table is the only record
-                        // that it happened.
+                        // sqlx has already reconnected and re-subscribed; only the gap needs
+                        // closing. The table is the only record that anything notified during the
+                        // drop happened.
                         //
-                        // This makes delivery at-least-once rather than exactly-once: a request
-                        // notified just before the drop can still be sitting in sqlx's buffer, so
-                        // the rescan replays it and the buffered notification then dispatches it
-                        // again. Observed, 2ms apart. That is the accepted shape of this design —
-                        // every `Recovery::Replay` command is idempotent because replay *is* the
-                        // recovery mechanism, and the only cost is a duplicate `_completed` row in
-                        // the audit trail. Suppressing it would mean a round trip per dispatch to
-                        // ask a question the in-flight claim exists to avoid asking.
+                        // Delivery is therefore at-least-once: a request notified just before the
+                        // drop can still be in sqlx's buffer, so the rescan replays it and the
+                        // buffer dispatches it again. Every `Recovery::Replay` command is
+                        // idempotent because replay *is* the recovery mechanism, and the cost is a
+                        // duplicate `_completed` row.
                         warn!("Listener connection was re-established; rescanning for missed requests");
                         handlers::recover(Arc::clone(&state)).await;
                     }
