+++ src/portfolio.rs
@@ -1,12 +1,8 @@
 //! The strategy: which pairs to hold, how large, and when to let them go.
 //!
-//! [`pairs`] is the record — the long/short mapping and the signal that justified it. [`screen`]
-//! decides what is worth holding, [`size`] how much, and [`risk`] whether the book can take it.
-//! [`execute`] is the only module that sends an order, and [`evaluate`] is the five-minute pass
-//! that drives the rest.
-//!
-//! [`account`] is the post-close half: what Alpaca says actually happened, written back as the
-//! reference the next session's drawdown gate reads.
+//! [`execute`] is the only module that sends an order; [`evaluate`] is the five-minute pass that
+//! drives the rest. [`account`] is the post-close half, writing back what Alpaca says actually
+//! happened as the reference the next session's drawdown gate reads.
 
 pub mod account;
 pub mod evaluate;
